requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
APScheduler==3.10.4
python-dotenv==1.0.1
gunicorn==22.0.0